Move GCM-specific garden bootstrapper into `app/bootstrappers`
In line with other bootstrapping-specific code for other binaries
