+++ cmd/gardener-controller-manager/app/bootstrappers/bootstrap_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package garden
+package bootstrappers
 
 import (
 	"context"
